Refactor uninstallation commands into subdirectory
